Bring us into the realm of PRODUCTION
This is a massive overhaul of the UI inner workings.
Most particularly, two of the Fragments in the UI were operating incredibly slowly. This is because they were initially written in a "quick and dirty" manner. Those two Fragments were the Files and Logs Fragments. Basically, they were loading *everything* and then dumping it all on the screen in a ScrollView. This, of course, takes a long time, and is very memory hungry.

So swapped those over to RecyclerView's.

Also moved logging (both ffmpeg log AND gps log) into sqlite3 database, and altered the Log Fragment to use a custom database backed Recycler Adapter, which means that it will only read log lines as they are displayed, rather than the whole thing at once.

The REAPER has also been updated to clean logs that are at least as old as the last (newest) file to be deleted.

Two settings added to manage logging;
1) Internal Log: Enable this to store the log on the INTERNAL STORAGE. This can be useful in deb