Fixes issue #15
If you enable the "Auto Share" switch in the settings screen, then when "checking" a file in the files screen, it will automatically launch a send-to dialog. The selection in this dialog will be remembered for next time, but only for this application. To clear the setting, disable and re-enable auto-share in the settings screen.