For the daring
This has a hand patched kernel that should disable the USB OTG whitelist feature.

If you feel daring, try installing this boot.img to your joying intel sofia android 6.0 device, and cross your fingers that it boots. I don't have the luxury of testing this.

If it doesn't boot, should be possible to reinstall the 6.0 update by renaming the update file from ...20... to ...60...
If it DOES boot, please run "cat /proc/kmsg" and capture the output while plugging in a UVC camera.
Also check the output of "lsusb" before and after,
and check if it creates a device file at /dev/video5.